Consolidate class and metaclass property pages
So class related info is properly shown again.
