+++ gaphor/diagram/profiles/__init__.py
@@ -4,7 +4,6 @@ from gaphor.diagram.profiles.extension import ExtensionItem
 def _load():
     from gaphor.diagram.profiles import (
         extensionconnect,
-        metaclasspropertypage,
         stereotypepage,
     )
